Update pipeline and structure
