+++ .eslintrc.js
@@ -2,6 +2,7 @@ module.exports = {
   env: {
     es6: true,
     node: true,
+    jest: true,
   },
   parserOptions: {
     "ecmaVersion": 2018,
