+++ .github/workflows/firebase-publish.yml
@@ -1,22 +1,29 @@
-name: Deploy to Firebase Functions via github action
-
+name: Deploy to Firebase Functions
+run-name: ${{ github.actor }} pushed to branch and triggered deploy
 on:
   push:
-    branches:
-      - main
+    branches: ["main" , "feature/*"]
   workflow_dispatch:
 
 jobs:
+  integration:
+    uses: ./.github/workflows/node-ci-workflow.yml
+
+  
   build_and_deploy:
+    needs: [integration]
+
     runs-on: ubuntu-22.04
     timeout-minutes: 5
+
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
+      - name: Set up Node v18
+        uses: actions/setup-node@v3
         with:
-          node-version: '16'
+          node-version: '18'
 
-      - name: Download dependencies
+      - name: Install dependencies
         working-directory: .
         run: find ./* -maxdepth 1 -name package.json -execdir npm ci \;
