+++ firebase.json
@@ -1,7 +1,7 @@
 {
   "functions": [
     {
-      "source": "update-dashboard",
+      "source": ".",
       "codebase": "update-dashboard",
       "ignore": [
         "node_modules",
@@ -11,16 +11,6 @@
       ],
       "predeploy": [
         "npm --prefix \"$RESOURCE_DIR\" run lint"
-      ]
-    },
-    {
-      "source": "update-dashboard-2",
-      "codebase": "update-dashboard-2",
-      "ignore": [
-        "node_modules",
-        ".git",
-        "firebase-debug.log",
-        "firebase-debug.*.log"
       ]
     }
   ],
