+++ junit.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<testsuites name="jest tests" tests="1" failures="0" errors="0" time="0.249">
+  <testsuite name="Update Dashboard TestCases" errors="0" failures="0" skipped="0" timestamp="2023-11-09T10:21:58" time="0.231" tests="1">
+    <testcase classname="Update Dashboard TestCases Check Event Type" name="Update Dashboard TestCases Check Event Type" time="0.014">
+    </testcase>
+  </testsuite>
+</testsuites>
