+++ package.json
@@ -1,12 +1,12 @@
 {
-  "name": "update-dashboard-2",
+  "name": "update-dashboard",
   "description": "Cloud Functions for Firebase",
   "scripts": {
     "lint": "eslint .",
     "serve": "firebase emulators:start --only functions",
     "shell": "firebase functions:shell",
     "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
+    "deploy": "find ./* -maxdepth 1 -name package.json -execdir npm ci \\; | npx firebase-tools deploy --only functions --project $FIREBASE_PROJECT",
     "logs": "firebase functions:log"
   },
   "engines": {
@@ -15,12 +15,14 @@
   "main": "index.js",
   "dependencies": {
     "firebase-admin": "^11.8.0",
-    "firebase-functions": "^4.3.1"
+    "firebase-functions": "^4.5.0"
   },
   "devDependencies": {
     "eslint": "^8.15.0",
     "eslint-config-google": "^0.14.0",
-    "firebase-functions-test": "^3.1.0"
+    "firebase-functions-test": "^3.1.0",
+    "jest": "^29.7.0",
+    "jest-junit": "^16.0.0"
   },
   "private": false
 }
