+++ test/index.test.js
@@ -1,0 +1,7 @@
+const {checkEventType} = require("../index");
+
+describe("Update Dashboard TestCases", () => {
+  test("Check Event Type", () => {
+    expect(checkEventType({"before": {"exists": true}, "after": {"exists": false}})).toBe("delete");
+  });
+});
